Allow h3:parent to return the parent at the current resolution
Before this change, asking for the parent of an h3 index at the same
resolution threw an error. Instead it's more useful to allow it to
return the same index as was passed in (eg the parent of an index with
resolution 10 at resoltion 10 is the same as the index input). This
change relaxes the check and adds tests for it.